% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_ctd.R
\name{append_ctd}
\alias{append_ctd}
\title{append CTD data to metadata}
\usage{
append_ctd(cids, smru, meta)
}
\arguments{
\item{cids}{SMRU campaign id(s)}

\item{smru}{SMRU table file - output of \code{pull_smru_tables}}

\item{meta}{clean metadata object - output of \code{clean_meta}}
}
\description{
calculates CTD end dates for each id and appends to corresponding metadata record. This is used to right-truncate deployments that fail the first 1-2 SSM-filtering rounds - usually due to sporadic tag function toward end of deployment. The tags usually stop recording (or transmitting?) CTD profiles prior to extreme 'sporadicity'.
}
\examples{

}
