% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_wc_DSA.R
\name{append_wc_DSA}
\alias{append_wc_DSA}
\title{Interpolate & append SSM locations to WC DSA datafiles}
\usage{
append_wc_DSA(wc, locs, deploy_meta, dropIDs)
}
\arguments{
\item{wc}{the WC DSA datafiles list}

\item{locs}{SSM location estimates}

\item{deploy_meta}{the deployment metadata object created in `append_wc()`}

\item{dropIDs}{DeploymentID's to be ignored}
}
\description{
interpolate SSM locations to WC DSA datafiles event times &
append to WC DSA datafiles
}
\keyword{internal}
