% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_wc_ECDHistosDSA.R
\name{append_wc_ECDHistosDSA}
\alias{append_wc_ECDHistosDSA}
\title{Interpolate & append SSM locations to WC ECDHistos_SCOUT_DSA datafiles}
\usage{
append_wc_ECDHistosDSA(wc, locs, deploy_meta, dropIDs)
}
\arguments{
\item{wc}{the WC ECDHistos_SCOUT_DSA datafiles list}

\item{locs}{SSM location estimates}

\item{deploy_meta}{the deployment metadata object created in `append_wc()`}

\item{dropIDs}{DeploymentID's to be ignored}
}
\description{
interpolate SSM locations to WC ECDHistos_SCOUT_DSA datafiles event times &
append to WC ECDHistos_SCOUT_DSA datafiles
}
\keyword{internal}
