% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_wc_Locations.R
\name{append_wc_Locations}
\alias{append_wc_Locations}
\title{Interpolate & append SSM locations to WC Locations file}
\usage{
append_wc_Locations(wc, locs, deploy_meta, dropIDs)
}
\arguments{
\item{wc}{the wc tag datafile list}

\item{locs}{SSM location estimates}

\item{deploy_meta}{the deployment metadata object created in `append_wc()`}

\item{dropIDs}{DeploymentID's to be ignored}
}
\description{
interpolate SSM locations to Locations event times &
append to Locations WC tag datafile
}
\keyword{internal}
