% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_ssm.R
\name{approx_ssm}
\alias{approx_ssm}
\title{general interpolation function}
\usage{
approx_ssm(sloc, wc)
}
\arguments{
\item{sloc}{data.frame of SSM estimated locations}

\item{wc}{the wc tag datafile list}
}
\description{
interpolate SSM locations to WC tag datafile event times
}
\keyword{internal}
