% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atn_wc_qc.R
\name{atn_wc_qc}
\alias{atn_wc_qc}
\title{ATN WC tag QC workflow}
\usage{
atn_wc_qc(
  wd = NULL,
  datadir = NULL,
  meta.file = NULL,
  outdir = NULL,
  collab.id = NULL,
  wc.akey = NULL,
  wc.skey = NULL,
  dropIDs = NULL,
  model = "rw",
  vmax = 3,
  time.step = 6,
  proj = NULL,
  reroute = FALSE,
  dist = 1000,
  buffer = 0.5,
  centroids = TRUE,
  cut = FALSE,
  min.gap = 72,
  QCmode = "nrt",
  p.int = 6,
  output = FALSE,
  ...
)
}
\arguments{
\item{wd}{the path to the working directory that contains: 1) the data directory
where tag data files are stored (if source = \code{local}); 2) the metadata directory
where all metadata files are stored; and 3) the destination directory for QC output.}

\item{datadir}{the name of the data directory (to be added to the \code{wd} path).}

\item{meta.file}{the metadata filename. Must reside within the \code{wd}}

\item{outdir}{the name of the QC output directory where CSV files will be
written (to be added to the \code{wd} path).}

\item{collab.id}{the Wildlife Computers ID for a collaborator sharing data}

\item{wc.akey}{a user's Access Key issued by Wildlife Computers for their API}

\item{wc.skey}{a user's Secret Key issued by Wildlife Computers for their API}

\item{dropIDs}{the ATN DeploymentID's that are to be ignored during the QC process}

\item{model}{the aniMotum SSM model to be used for the location QC - typically
either \code{rw} or \code{crw}.}

\item{vmax}{for SSM fitting; max travel rate (m/s) to identify implausible locations}

\item{time.step}{the prediction interval (in hours) to be used by the SSM}

\item{proj}{the proj4string to be used for the location data & for the SSM-estimated
locations. The default (NULL) will result in one of 3 projections being used,
depending on whether the centroid of the observed latitudes lie in N or S polar
regions, temperate or equatorial regions, or if tracks straddle (or lie close to)
-180,180 longitude.}

\item{reroute}{whether QC'd tracks should be re-routed off of land
(default is FALSE). Note, in some circumstances this can substantially increase
processing time. Default land polygon data are sourced from the
\code{ropensci/rnaturalearthhires} R package.}

\item{dist}{the distance in km from outside the convex hull of observed
locations from which to select land polygon data for re-routing. Ignored if
\code{reroute = FALSE}.}

\item{buffer}{the distance in km to buffer rerouted locations from the
coastline. Ignored if \code{reroute = FALSE}.}

\item{centroids}{whether centroids are to be included in the visibility graph
mesh used by the rerouting algorithm. See \code{?pathroutr::prt_visgraph} for
details. Ignored if \code{reroute = FALSE}.}

\item{cut}{logical; should predicted locations be dropped if they lie within
in a large data gap (default is FALSE).}

\item{min.gap}{the minimum data gap duration (h) to be used for cutting
predicted locations (default is 72 h)}

\item{QCmode}{one of either \code{nrt} for Near Real-Time QC or \code{dm} for Delayed
Mode QC.}

\item{p.int}{prediction interval to use for sub-sampling predicted locations
(default = 6 h)}

\item{output}{logical; should fn return a list of QC-generated objects. This
results in a single large object return that can be useful for troubleshooting
QC errors or undesirable results.}

\item{...}{additional arguments to be passed to SSM model fitting, see
\code{aniMotum::fit_ssm} & \code{aniMotum::route_path} for details.}
}
\description{
Wrapper function that executes the complete workflow from data
download to SSM-appended tag data files output as CSV files.
}
