% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_diag.R
\name{clean_diag}
\alias{clean_diag}
\title{clean diag files}
\usage{
clean_diag(smru)
}
\arguments{
\item{smru}{list of SMRU tables}
}
\description{
restructures diag files, formats dates & lc's in preparation for SSM-filtering
}
\examples{

}
