% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_meta.R
\name{clean_meta}
\alias{clean_meta}
\title{clean metadata}
\usage{
clean_meta(cids,
  file = "~/Dropbox/collab/imos/qc/r/meta/IMOS_CTD_metadata_11042019.csv")
}
\arguments{
\item{cids}{SMRU campaign ids}

\item{file}{path to metadata .csv file}
}
\description{
restructures metadata, formats dates, standardizes variable names, subsets to specified campaign ids
}
\examples{

}
