% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download SMRU or Wildlife Computers tag data from their servers}
\usage{
download_data(
  dest = NULL,
  source = "smru",
  cid = NULL,
  user = NULL,
  pwd = NULL,
  wc.akey = "VVeerW+G6YUe7olzlrOr6q5o2Nkjx5PTEwuwElcPyKc=",
  wc.skey = "7k9MupziDacYNur/3IPMDjn7wum6oQk5eV2LRj86iDw=",
  owner.id = NULL,
  subset.ids = NULL,
  download = TRUE,
  ...
)
}
\arguments{
\item{dest}{destination path to save download}

\item{source}{source type of data to be downloaded. Can be one of:
\itemize{
\item \code{smru} - SMRU Data Server;
\item \code{wc} - Wildlife Computers Data Portal API;
}}

\item{cid}{SMRU tag deployment campaign id(s) to download, eg. "ct180"}

\item{user}{SMRU data server username as a quoted string}

\item{pwd}{SMRU data server password as a quoted string}

\item{wc.akey}{an Access Key issued by Wildlife Computers for their API}

\item{wc.skey}{a Secret Key issued by Wildlife Computers for their API}

\item{owner.id}{the Wildlife Computers uuid associated with the data owner}

\item{subset.ids}{a single column .CSV file of WC UUID's to be included in
the QC, with uuid as the variable name.}

\item{download}{(logical) indicating if the data is to be downloaded from
the tag manufacturer's server. If the source is \code{wc} and \code{download = FALSE}
then only the Wildlife Computers tag deployment metadata is downloaded.}

\item{...}{additional arguments passed to \code{smru_get_mdb} or \code{wc_get_files()}}
}
\description{
Satellite tracking data are accessed from the SMRU data server,
or accessed from the Wildlife Computers Portal API via the \code{source} argument.
}
