% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metadata.R
\name{download_meta}
\alias{download_meta}
\title{download deployment metadata}
\usage{
download_meta(
  source = "smru",
  tag_mfr = "smru",
  tag_data = NULL,
  cids = NULL,
  dropIDs = NULL,
  file = NULL
)
}
\arguments{
\item{source}{the source of the metadata, current options are `imos`, `smru`,
`atn`.}

\item{tag_mfr}{the tag manufacturer, current options are `smru` or `wc`
cids SMRU campaign ids}

\item{tag_data}{a list of either `smru` data tables or `wc` files as output by
`pull_data`.}

\item{dropIDs}{SMRU refs or WC ids to be dropped}

\item{file}{path to metadata .csv file, if provided then metadata will be
downloaded from the provided `source`}
}
\description{
downloads, restructures & formats metadata, appends
dive/CTD start and end datetimes (for QC), & fills in missing required
metadata - eg. release_datetime, release_longitude/latitude's with data from
the GPS (if present) or Argos location file.
}
