% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_QC.R
\name{grab_QC}
\alias{grab_QC}
\title{grab \code{tibble}'s by name from a \code{aniMotum} model object}
\usage{
grab_QC(
  x,
  what = "fitted",
  cut = FALSE,
  as_sf = FALSE,
  normalise = FALSE,
  group = FALSE
)
}
\arguments{
\item{x}{a \code{aniMotum} \code{ssm_df} or \code{mpm_df} model object}

\item{what}{the tibble to be grabbed; either \code{fitted}, \code{predicted},
\code{rerouted} (\code{ssm_df} only), or \code{data} (single letters can be used).}

\item{cut}{logical; if TRUE then drop any predicted locations associated with a
\code{keep = FALSE} variable - ie. locations occurring within a data gap larger
than some user-specified threshold duration. default = FALSE}

\item{as_sf}{logical; if FALSE (default) then return a \code{tibble} with
un-projected longlat coordinates, otherwise return an \verb{sf tibble}. Ignored
if x is an \code{mpm} model object.}

\item{normalise}{logical; if output includes a move persistence estimate,
should \code{g} (the move persistence index) be normalised to have minimum = 0 and
maximum = 1 (default = FALSE). Note, this normalisation is not applied to the
standard errors of the logit-scale move persistence estimates (\code{logit_g},
\code{logit_g.se}).}

\item{group}{logical; should \code{g} be normalised among individuals as a group,
a 'relative g', or to individuals separately to highlight regions of lowest
and highest move persistence along single tracks (default = FALSE).}
}
\value{
a \code{tibble} with all individual \code{tibble}'s appended
}
\description{
\code{grab_QC()} is the \code{ArgosQC} version of \code{aniMotum::grab()}. It lets
you obtain \code{fitted}, \code{predicted}, \code{rerouted} or \code{data} \code{tibble}'s from a
compound \code{tibble} created when fitting to multiple individual data sets. The
specified \code{tibble}'s are appended to a single output \code{tibble}. Unlike
\code{aniMotum::grab()}, \code{ArgosQC::grab_QC()} respects additional columns in the
predicted locations \code{tibble}, allowing a logical vector, \code{keep}, to be used
to delineate data gaps that may be associated with spurious predicted locations.
}
\details{
if multiple \code{ssm_df} model objects are present in \code{x}, \code{as_sf = TRUE},
and at least 1 estimated track has a coordinate reference system (\code{crs}) with
longitude centered on 180 (e.g. a track straddling -180,180) then all tracks
will be re-projected to that \code{crs}.
}
\examples{
## generate an ssm fit object
xs <- aniMotum::fit_ssm(ellie, model = "rw", time.step=24,
control = ssm_control(verbose = 0))

## grab predicted values as an un-projected tibble
preds <- grab_QC(xs, what = "predicted", cut = TRUE)

}
