% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imos_smru_qc.R
\name{imos_smru_qc}
\alias{imos_smru_qc}
\title{IMOS SMRU SRDL QC workflow}
\usage{
imos_smru_qc(wd, config)
}
\arguments{
\item{config}{a JSON configuration file containing the following parameters:
\itemize{
\item \code{datadir} the name of the data directory (to be added to the \code{wd} path).
\item \code{meta.file} the metadata filename. Must reside within the \code{wd}.
\item \code{maps.dir} the directory path to write diagnostic maps of QC'd tracks.
\item \code{diag.dir} the directory path to write diagnostic time-series plots of QC'd lon & lat.
\item \code{outdir} the name of the QC output directory where CSV files will be
written (to be added to the \code{wd} path).
\item \code{cid} SMRU campaign (AODN sattag_program) ID.
\item \code{sp} the species being QC'd - currently, either "sese" or "ortu".
\item \code{smru.usr} SMRU data server username as a string.
\item \code{smru.pwd} SMRU data server password as a string.
\item \code{dwnld.timeout} extends the download timeout period a specified number of seconds.
\item \code{dropIDs} the SMRU ref ID's that are to be ignored during the QC process.
\item \code{proj} the proj4string to be used for the location data & for the SSM-estimated
locations. The default (NULL) will result in one of 3 projections being used,
depending on whether the centroid of the observed latitudes lie in N or S polar
regions, temperate or equatorial regions, or if tracks straddle (or lie close to)
-180,180 longitude.
\item \code{model} the aniMotum SSM model to be used for the location QC - typically
either \code{rw} or \code{crw}.
\item \code{vmax} for SSM fitting; max travel rate (m/s) to identify implausible locations
\item \code{time.step} the prediction interval (in hours) to be used by the SSM
\item \code{reroute} whether QC'd tracks should be re-routed off of land
(default is FALSE). Note, in some circumstances this can substantially increase
processing time. Default land polygon data are sourced from the
\code{ropensci/rnaturalearthhires} R package.
\item \code{dist} the distance in km from outside the convex hull of observed
locations from which to select land polygon data for re-routing. Ignored if
\code{reroute = FALSE}.
\item \code{buffer} the distance in km to buffer rerouted locations from the
coastline. Ignored if \code{reroute = FALSE}.
\item \code{centroids} whether centroids are to be included in the visibility graph
mesh used by the rerouting algorithm. See \code{?pathroutr::prt_visgraph} for
details. Ignored if \code{reroute = FALSE}.
\item \code{cut} logical; should predicted locations be dropped if they lie within
in a large data gap (default is FALSE).
\item \code{min.gap} the minimum data gap duration (h) to be used for cutting
predicted locations (default is 72 h)
\item \code{QCmode} one of either \code{nrt} for Near Real-Time QC or \code{dm} for Delayed
Mode QC.
\item \code{output} logical; should fn return a list of QC-generated objects. This
results in a single large object return that can be useful for troubleshooting
QC errors or undesirable results.
\item \code{dwnld} logical; should the tag data be downloaded from the SMRU server
(default TRUE). Useful when testing so unnecessary load is not put on the
SMRU server.
}}

\item{`wd`}{the path to the working directory that contains: 1) the data directory
where tag data files are stored (if source = \code{local}); 2) the metadata directory
where all metadata files are stored; and 3) the destination directory for QC output.}
}
\description{
Wrapper function that executes the complete workflow from data
download to SSM-appended tag data files output as CSV files.
}
