% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_wc_locations.R
\name{prep_wc}
\alias{prep_wc}
\title{clean, truncate & convert WC Locations files to sf-tibbles}
\usage{
prep_wc(
  wc,
  meta,
  dropIDs,
  crs = "+proj=merc +units=km +ellps=WGS84 +no_defs",
  QCmode = "nrt"
)
}
\arguments{
\item{wc}{list of WC datafiles}

\item{meta}{metadata used to truncate start of diag data for each individual}

\item{dropIDs}{WC DeploymentID's to be dropped (eg. tags were turned on but not deployed)}

\item{crs}{a proj4string to re-project diag locations from longlat}

\item{QCmode}{specify whether QC is near real-time (nrt) or delayed-mode (dm),
in latter case diag is not right-truncated & date of first dive is used
for the track start date}
}
\description{
restructures Locations files, formats dates & lc's; truncates start
(and end for "nrt") of individual deployments using ctd dates; converts to
sf geometry - all in preparation for SSM-filtering. Splits resulting truncated
diag files by species.
}
