% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_2_aodn.R
\name{push_2_aodn}
\alias{push_2_aodn}
\title{Zip QC'd .csv files for transfer to AODN}
\usage{
push_2_aodn(
  cids,
  path = NULL,
  user = NULL,
  host = NULL,
  dest = NULL,
  pwd = NULL,
  nopush = TRUE,
  suffix = "_nrt"
)
}
\arguments{
\item{cids}{campaign ids to zip}

\item{path}{path to the QC'd .csv files}

\item{user}{no longer used (formerly the rsync username); ignored}

\item{host}{no longer used (formerly the rsync server); ignored}

\item{dest}{no longer used (formerly the rsync destination); ignored}

\item{pwd}{no longer used (formerly the rsync password); ignored}

\item{nopush}{must be \code{TRUE}; transfer to AODN is done by sftp outside this
function}

\item{suffix}{suffix to add to zip files (_nrt or _dm)}
}
\description{
Zips the QC'd .csv files of each campaign into
\verb{<cid><suffix>.zip} in \code{path}, then deletes the .csv files. The zip files are
transferred to AODN by sftp outside this function.
}
\keyword{internal}
