% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_multi_filter.R
\name{redo_multi_filter}
\alias{redo_multi_filter}
\title{redo failed multi-filter cases}
\usage{
redo_multi_filter(fit)
}
\arguments{
\item{fit}{foieGras fit object from first round of filtering}

\item{diag_sf}{\code{sf}-projected diag to be used}

\item{meta}{metadata to split diag by species}
}
\description{
re-apply SSM filter to diag data for id's that failed to converge. parallelized
}
\examples{

}
