% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_append_ssm.R
\name{smru_append_ssm}
\alias{smru_append_ssm}
\title{append SMRU tables with SSM-estimated locations}
\usage{
smru_append_ssm(
  smru,
  fit,
  what = "p",
  meta,
  gps.tab = FALSE,
  cut = FALSE,
  dropIDs = NULL
)
}
\arguments{
\item{smru}{SMRU table file - output of \code{pull_smru_tables}}

\item{fit}{final \code{foieGras} fit object}

\item{what}{choose which locations to use for annotating SMRU tables (default = "predicted")}

\item{meta}{metadata used to truncate start of diag data for each individual}

\item{gps.tab}{logical; does a SMRU GPS table exist, if so append to this as well}

\item{cut}{drop predicted locations if keep = FALSE, ie. locations in a large data gap}

\item{dropIDs}{SMRU refs to be dropped}
}
\description{
append SMRU tables so each event has SSM-derived lon, lat, x, y, x.se, y.se.
}
