% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_build_meta_imos.R
\name{smru_build_meta_imos}
\alias{smru_build_meta_imos}
\title{construct IMOS metadata from SMRU data server & config file details}
\usage{
smru_build_meta_imos(
  cid,
  user = NULL,
  pwd = NULL,
  dropIDs,
  meta.args = NULL,
  tag_data
)
}
\arguments{
\item{cid}{the SMRU campaign ID - supplied by \code{conf$harvest$cid}}

\item{user}{SMRU data server username as a quoted string}

\item{pwd}{SMRU data server password as a quoted string}

\item{dropIDs}{SMRU refs or WC ids to be dropped from QC - supplied by
\code{conf$harvest$dropIDs} in \code{get_metadata()}}

\item{meta.args}{metadata fields to be passed from config file - required when
downloading tag metadata from SMRU server.}

\item{tag_data}{a list of SMRU data tables as output by \code{pull_data}.}
}
\description{
utility fn for constructing IMOS metadata when a .CSV metadata
file is not available. Required deployment info is scraped from the SMRU
Data server table associated with the provided \code{cid} (SMRU campaign id).
}
\keyword{internal}
