% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_clean_diag.R
\name{smru_clean_diag}
\alias{smru_clean_diag}
\title{clean diag files}
\usage{
smru_clean_diag(smru, dropIDs = NULL)
}
\arguments{
\item{smru}{list of SMRU tables}

\item{dropIDs}{SMRU refs to be dropped (eg. tags were turned on but not deployed)}
}
\description{
restructures diag files, formats dates & lc's in preparation for SSM-filtering
}
