% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_clean_imos_meta.R
\name{smru_clean_imos_meta}
\alias{smru_clean_imos_meta}
\title{clean metadata}
\usage{
smru_clean_imos_meta(cid, smru, dropIDs = NULL, file = NULL)
}
\arguments{
\item{cid}{SMRU campaign ids}

\item{smru}{SMRU table}

\item{dropIDs}{SMRU refs to be dropped}

\item{file}{path to metadata .csv file}
}
\description{
restructures metadata, formats dates,
standardizes variable names, subsets to specified campaign ids, appends
dive start and end datetimes (for QC), & fills in missing
release_longitude/latitude's with the first best quality (lc 3,2, or 1)
Argos location after release_datetime
}
\keyword{internal}
