% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_clean_meta_atn.R
\name{smru_clean_meta_atn}
\alias{smru_clean_meta_atn}
\title{read & handle ATN metadata from .CSV file}
\usage{
smru_clean_meta_atn(
  file,
  col.types.string = c("iiccccccTccicccccTTcddcccicdccdccdcccdcTc"),
  cid,
  dropIDs
)
}
\arguments{
\item{file}{filepath to metadata .csv file - supplied by \code{get_metadata()}}

\item{col.types.string}{string of single character col_types for \code{col_types}
arg in \code{read_csv}}

\item{cid}{the SMRU campaign ID - supplied by \code{conf$harvest$cid}}

\item{dropIDs}{SMRU refs or WC ids to be dropped from QC - supplied by
\code{conf$harvest$dropIDs} in \code{get_metadata()}}
}
\description{
utility fn for handling fully specified ATN SMRU tag metadata
}
\keyword{internal}
