% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_ctd_write.R
\name{smru_ctd_write}
\alias{smru_ctd_write}
\title{SMRU ctd table tests & write to .csv}
\usage{
smru_ctd_write(
  smru_ssm,
  meta,
  program = "imos",
  path = NULL,
  drop.refs = NULL,
  suffix = "_nrt"
)
}
\arguments{
\item{smru_ssm}{SSM-appended SMRU table file - output of \code{append_ssm}}

\item{meta}{metadata}

\item{program}{Determines structure of output metadata. Currently, either `imos` or `atn`.}

\item{path}{path to write .csv files}

\item{drop.refs}{individual ids to be dropped}

\item{suffix}{suffix to add to .csv files (_nrt, _dm, or _hist)}
}
\description{
Apply AODN tests to SMRU ctd table, write to .csv - format depends on program (IMOS, ATN)
}
