% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_get_mdb.R
\name{smru_get_mdb}
\alias{smru_get_mdb}
\title{Download .mdb files from the SMRU server}
\usage{
smru_get_mdb(
  cid,
  dest = NULL,
  user = NULL,
  pwd = NULL,
  timeout = 120,
  verbose = FALSE
)
}
\arguments{
\item{cid}{SMRU campaign ids to be downloaded from SMRU data server}

\item{dest}{destination path for saving .mdb files}

\item{user}{SMRU data server username as a string}

\item{pwd}{SMRU data server password as a string}

\item{timeout}{timeout duration in s (default 120 s); sets
options(timeout = timeout) to avoid mdb.zip download error from SMRU server}

\item{verbose}{turn on/off progress indicator}
}
\description{
fetches .mdb files from SMRU server and saves to a \code{dest}-ination directory
}
\keyword{internal}
