% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_pull_tables.R
\name{smru_pull_tables}
\alias{smru_pull_tables}
\title{Pull tables from SMRU .mdb files}
\usage{
smru_pull_tables(
  cids,
  path2mdb,
  tables = c("diag", "gps", "haulout", "ctd", "dive", "cruise", "summary"),
  p2mdbtools = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cids}{SMRU campaign ids}

\item{path2mdb}{path to SMRU .mdb file(s)}

\item{tables}{specify which tables to extract, default is to extract all tables}

\item{p2mdbtools}{path to mdbtools binaries. Specifying the path can avoid
an error when calling from within RStudio, eg. on MacBook Pro M1 Pro with
homebrew-installed mdbtools @ /opt/homebrew/Cellar/mdbtools/1.0.0/bin/}

\item{verbose}{turn on/off progress indicator}
}
\description{
extracts specified tables from SMRU .mdb files, using Hmisc::mdb.get
}
