% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_ssm_outputs.R
\name{smru_ssm_outputs}
\alias{smru_ssm_outputs}
\title{create ssmoutputs data.frame}
\usage{
smru_ssm_outputs(fit, what, dropIDs, suffix)
}
\arguments{
\item{fit}{final \code{aniMotum} fit object}

\item{what}{specify whether predicted or rerouted locations are to be used}

\item{dropIDs}{individual SMRU ids to be dropped}

\item{suffix}{suffix to add to .csv files (_nrt, _dm, or _hist)}
}
\description{
build ssmoutputs data.frame from SSM predicted location
}
\keyword{internal}
