% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_truncate_ssm.R
\name{smru_truncate_ssm}
\alias{smru_truncate_ssm}
\title{right-truncate all SSM-appended SMRU tables using CTD end date for given individuals}
\usage{
smru_truncate_ssm(smru_ssm, meta, refs)
}
\arguments{
\item{smru_ssm}{SSM-appended SMRU file to use}

\item{meta}{metadata used to truncate SSM-appended SMRU tables for each individual}

\item{refs}{device_id's (SMRU ref's) to apply truncation}
}
\description{
right-truncate all SSM-appended SMRU tables using CTD end date for given individuals, using CTD date-times from metadata
}
