% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_write_csv.R
\name{smru_write_cruise}
\alias{smru_write_cruise}
\title{SMRU cruise table - write to .csv}
\usage{
smru_write_cruise(
  smru_ssm,
  meta,
  program = "atn",
  path = NULL,
  dropIDs = NULL,
  suffix = "_nrt"
)
}
\arguments{
\item{smru_ssm}{SSM-appended SMRU table file - output of \code{append_ssm}}

\item{meta}{metadata}

\item{program}{Determines structure of output metadata. Currently, either
\code{imos} or \code{atn}. If \code{imos} then function returns nothing as cruise table is
not currently accepted by AODN.}

\item{path}{path to write .csv files}

\item{dropIDs}{individual ids to be dropped}

\item{suffix}{suffix to add to .csv files (_nrt, _dm, or _hist)}
}
\description{
Write SMRU cruise table to .csv - non-IMOS only
}
\keyword{internal}
