% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_mark_gaps.R
\name{ssm_mark_gaps}
\alias{ssm_mark_gaps}
\title{Identify SSM predicted & rerouted (if present) location estimates in data gaps}
\usage{
ssm_mark_gaps(ssm, min.gap = 24, mark = TRUE)
}
\arguments{
\item{ssm}{the SSM fit object from \code{redo_multi_filter()}}

\item{min.gap}{the minimum data gap duration from which SSM estimates are
removed (in hours)}

\item{mark}{logical; should the SSM data be marked (TRUE; default), otherwise
the function does no marking and returns the original SSM fit object}
}
\description{
Identify & mark SSM predicted & rerouted (if present) location
estimates in track segments with data gaps of a specified minimum duration.
}
