% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_outputs.R
\name{ssm_outputs}
\alias{ssm_outputs}
\title{create ssmoutputs data.frame}
\usage{
ssm_outputs(fit, what, drop.refs)
}
\arguments{
\item{fit}{final \code{aniMotum} fit object}

\item{what}{specify whether predicted or rerouted locations are to be used}

\item{drop.refs}{individual SMRU ids to be dropped}
}
\description{
build ssmoutputs data.frame from SSM predicted location
}
\keyword{internal}
