% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{tidy_download}
\alias{tidy_download}
\title{download helper fn, borrowed from usethis:::tidy_download - without the interactivity}
\usage{
tidy_download(url, destdir = getwd())
}
\arguments{
\item{url}{url for data}

\item{destdir}{destination path to save download}
}
\description{
helper fn
}
\keyword{internal}
