% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_smru_ssm.R
\name{truncate_smru_ssm}
\alias{truncate_smru_ssm}
\title{right-truncate all SSM-appended SMRU tables using CTD end date for given individuals}
\usage{
truncate_smru_ssm(smru_ssm, fit, meta, refs)
}
\arguments{
\item{smru_ssm}{SSM-appended SMRU file to use}

\item{meta}{metadata used to truncate SSM-appended SMRU tables for each individual}

\item{ref}{device_id's (SMRU ref's) to apply truncation}
}
\description{
right-truncate all SSM-appended SMRU tables using CTD end date for given individuals, using CTD date-times from metadata
}
\examples{

}
