% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_append_ssm.R
\name{wc_append_datafile}
\alias{wc_append_datafile}
\title{Interpolate & append SSM locations to WC tag datafile}
\usage{
wc_append_datafile(wc, locs, deploy_meta, dropIDs)
}
\arguments{
\item{wc}{the WC tag datafile list}

\item{locs}{SSM location estimates}

\item{deploy_meta}{the deployment metadata object created in `append_wc()`}

\item{dropIDs}{DeploymentID's to be ignored}
}
\description{
interpolate SSM locations to WC tag datafile event times &
append to WC tag datafile
}
\keyword{internal}
