% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_append_ssm.R
\name{wc_append_ECDHistosDSA}
\alias{wc_append_ECDHistosDSA}
\title{Interpolate & append SSM locations to WC ECDHistos_SCOUT_DSA datafiles}
\usage{
wc_append_ECDHistosDSA(wc, locs, deploy_meta, dropIDs)
}
\arguments{
\item{wc}{the WC ECDHistos_SCOUT_DSA datafiles list}

\item{locs}{SSM location estimates}

\item{deploy_meta}{the deployment metadata object created in \code{append_wc()}}

\item{dropIDs}{DeploymentID's to be ignored}
}
\description{
interpolate SSM locations to WC ECDHistos_SCOUT_DSA datafiles event times &
append to WC ECDHistos_SCOUT_DSA datafiles
}
\keyword{internal}
