% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_append_ssm.R
\name{wc_append_Locations}
\alias{wc_append_Locations}
\title{Interpolate & append SSM locations to WC Locations file}
\usage{
wc_append_Locations(wc, locs, deploy_meta, dropIDs)
}
\arguments{
\item{wc}{the wc tag datafile list}

\item{locs}{SSM location estimates}

\item{deploy_meta}{the deployment metadata object created in \code{append_wc()}}

\item{dropIDs}{DeploymentID's to be ignored}
}
\description{
interpolate SSM locations to Locations event times &
append to Locations WC tag datafile
}
\keyword{internal}
