% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_append_ssm.R
\name{wc_append_ssm}
\alias{wc_append_ssm}
\title{append WC tag datafiles with SSM-estimated locations}
\usage{
wc_append_ssm(
  wc,
  fit,
  what = "p",
  meta,
  cut = FALSE,
  dropIDs = NULL,
  crs = "+proj=merc +units=km +ellps=WGS84 +no_defs"
)
}
\arguments{
\item{wc}{WC tag datafiles - output of \code{pull_wc}}

\item{fit}{final \code{aniMotum} fit object(s)}

\item{what}{choose which locations to use for annotating WC tag datafiles (default = "predicted")}

\item{meta}{metadata used to truncate start of diag data for each individual}

\item{cut}{drop predicted locations if keep = FALSE, ie. locations in a large data gap
(currently, only used in DM QC mode)}

\item{dropIDs}{SMRU DeploymentIDs to be dropped}

\item{crs}{CRS to be applied when interpolating SSM-estimated locations and re-projecting back from Cartesian coords to longlat}
}
\description{
append WC tag datafiles so each event has SSM-derived lon, lat, x, y, x.se, y.se.
}
