% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_build_meta_generic.R
\name{wc_build_meta_generic}
\alias{wc_build_meta_generic}
\title{construct generic deployment metadata from WC data portal xml & config
file details}
\usage{
wc_build_meta_generic(ids, dropIDs, meta.args, wc.meta, tag_data)
}
\arguments{
\item{ids}{subset of WC UUID's to consider - supplied by
\code{conf$harvest$tag.list} in \code{get_metadata()}}

\item{dropIDs}{WC ids to be dropped from QC - supplied by \code{conf$harvest$dropIDs}
in \code{get_metadata()}}

\item{meta.args}{metadata fields to be passed from config file - required when
downloading tag metadata from WC Data Portal API.}

\item{wc.meta}{an R data.frame of Wildlife Computers tag deployment metadata
obtained via \code{download_data()} - supplied by \code{get_metadata()}}

\item{tag_data}{a list of SMRU data tables as output by \code{pull_data}.}
}
\description{
utility fn for constructing generic metadata when a .CSV metadata
file is not available. Required deployment info is pulled from the xml file
download from the WC Data Portal API. XML file downloaded via \code{wc_get_files()}
}
\keyword{internal}
