% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_clean_meta_atn.r
\name{wc_clean_meta_atn}
\alias{wc_clean_meta_atn}
\title{read & handle ATN metadata from .CSV file}
\usage{
wc_clean_meta_atn(
  file,
  col.types.string = c("iiccccccTccicccccTTcddcccicdccdccdcccdcTc"),
  ids,
  dropIDs
)
}
\arguments{
\item{file}{filepath to metadata .csv file - supplied by \code{get_metadata()}}

\item{col.types.string}{string of single character col_types for \code{col_types}
arg in \code{read_csv}}

\item{ids}{subset of WC UUID's to consider - supplied by
\code{conf$harvest$tag.list} in \code{get_metadata()}}

\item{dropIDs}{SMRU refs or WC ids to be dropped from QC - supplied by
\code{get_metadata()} & provided by \code{conf$harvest$dropIDs}}
}
\description{
utility fn for handling fully specified ATN WC tag metadata
}
\keyword{internal}
