% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_clean_meta_irap.r
\name{wc_clean_meta_irap}
\alias{wc_clean_meta_irap}
\title{read & handle IRAP metadata from .CSV file & WC data portal}
\usage{
wc_clean_meta_irap(
  file,
  col.types.string = "cTcddcccdccdcccciciT",
  ids,
  dropIDs,
  wc.meta
)
}
\arguments{
\item{file}{filepath to metadata .csv file - supplied by \code{get_metadata()}}

\item{col.types.string}{string of single character col_types for \code{col_types}
arg in \code{read_csv}}

\item{ids}{subset of WC UUID's to consider - supplied by
\code{conf$harvest$tag.list} in \code{get_metadata()}}

\item{dropIDs}{SMRU refs or WC ids to be dropped from QC - supplied by
\code{conf$harvest$dropIDs} in \code{get_metadata()}}

\item{wc.meta}{an R data.frame of Wildlife Computers tag deployment metadata
obtained via \code{download_data()} - supplied by \code{get_metadata()}}
}
\description{
utility fn for organising IRAP (NSWOO turtle/seal) metadata
}
\keyword{internal}
