% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_get_collaborator_ids.R
\name{wc_get_collab_ids}
\alias{wc_get_collab_ids}
\title{WC API access to collaborator ids}
\usage{
wc_get_collab_ids(a.key = NULL, s.key = NULL, verbose = FALSE)
}
\arguments{
\item{a.key}{an Access Key issued by Wildlife Computers for their API}

\item{s.key}{a Secret Key issued by Wildlife Computers for their API}
}
\description{
Download collaborator ids via the Wildlife Computers Portal API.
User-supplied WC Access Key, WC Secret Key.
}
\details{
Returns a data.frame with WC ids and email addresses of collaborators
}
\keyword{internal}
