% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_get_uuids.R
\name{wc_get_uuids}
\alias{wc_get_uuids}
\title{WC API access to tag dataset UUID's}
\usage{
wc_get_uuids(a.key = NULL, s.key = NULL, owner.id = NULL, verbose = FALSE)
}
\arguments{
\item{a.key}{an Access Key issued by Wildlife Computers for their API}

\item{s.key}{a Secret Key issued by Wildlife Computers for their API}

\item{owner.id}{a WC data owner ID}
}
\description{
Download tag dataset UUID's via the Wildlife Computers Portal API.
User-supplied WC Access Key, WC Secret Key.
}
\details{
Returns a data.frame with WC tag dataset UUID's (id), data owner's
email address (owner), tag data status (status), tag type (tag),
& last update date of tag data (last_update_date)
}
\keyword{internal}
