% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_prep_locations.R
\name{wc_prep_loc}
\alias{wc_prep_loc}
\title{clean, truncate & convert WC Locations files to sf-tibbles}
\usage{
wc_prep_loc(wc, meta, dropIDs, crs = NULL, program = "atn", QCmode = "nrt")
}
\arguments{
\item{wc}{list of WC datafiles}

\item{meta}{metadata used to truncate start of diag data for each individual}

\item{dropIDs}{WC DeploymentID's to be dropped (eg. tags were turned on but not deployed)}

\item{crs}{a proj4string to re-project diag locations from longlat. Default is NULL
which results in one of 4 possible projections applied automatically, based on
the centroid of the tracks. See \code{overview} vignette for details.}

\item{program}{specify the aniBOS program contributing data (currently: 'atn', 'irap')}

\item{QCmode}{specify whether QC is near real-time (nrt) or delayed-mode (dm),
in latter case wc is not right-truncated & date of first dive is used
for the track start date.}
}
\description{
restructures Locations files, formats dates & lc's; truncates start
(and end for "nrt") of individual deployments using ctd dates; converts to
sf geometry - all in preparation for SSM-filtering. Splits resulting truncated
diag files by species.
}
