% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_ssm_outputs.R
\name{wc_ssm_outputs}
\alias{wc_ssm_outputs}
\title{create ssmoutputs data.frame}
\usage{
wc_ssm_outputs(fit, what, dropIDs, suffix, pred.int = 6)
}
\arguments{
\item{fit}{final \code{aniMotum} fit object}

\item{what}{specify whether predicted or rerouted locations are to be used}

\item{dropIDs}{individual WC DeploymentIDs to be dropped}

\item{suffix}{suffix to add to .csv files (_nrt, _dm, or _hist)}

\item{pred.int}{prediction interval to use for sub-sampling predicted locations
(default = 6 h)}
}
\description{
build ssmoutputs data.frame from SSM predicted/rerouted locations
}
\keyword{internal}
