% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_write_datafile.R
\name{wc_write_datafile}
\alias{wc_write_datafile}
\title{WC generic datafile write to .csv}
\usage{
wc_write_datafile(
  wc_ssm,
  meta,
  program = "atn",
  path = NULL,
  dropIDs = NULL,
  suffix = "_nrt"
)
}
\arguments{
\item{meta}{metadata}

\item{program}{Determines structure of output metadata. Currently, only `atn`.}

\item{path}{path to write .csv files}

\item{dropIDs}{individual ids to be dropped}

\item{suffix}{suffix to add to .csv files (_nrt, _dm, or _hist)}

\item{smru_ssm}{SSM-appended WC tag datafile - output of \code{append_ssm}}
}
\description{
Write WC generic datafile to .csv - format specific to ATN program
}
\keyword{internal}
