% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_write_ssm.R
\name{wc_write_ssm}
\alias{wc_write_ssm}
\title{ssmoutputs table write to .csv}
\usage{
wc_write_ssm(
  locs_out,
  meta,
  program = "atn",
  path = NULL,
  dropIDs = NULL,
  suffix = "_nrt"
)
}
\arguments{
\item{meta}{metadata}

\item{program}{Determines structure of output metadata. Currently, either `imos` or `atn`.}

\item{path}{path to write .csv files}

\item{dropIDs}{individual ids to be dropped}

\item{suffix}{suffix to add to .csv files (_nrt, _dm, or _hist)}

\item{smru_ssm}{SSM-appended WC tag datafile - output of \code{append_ssm}}
}
\description{
write to .csv - format depends on program (IMOS, ATN)
}
\keyword{internal}
