% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_2_csv.R
\name{write_2_csv}
\alias{write_2_csv}
\title{reconfigure annotated SMRU tables for AODN}
\usage{
write_2_csv(
  smru_ssm,
  fit,
  meta,
  path = "~/Dropbox/collab/imos/imos_qc/aodn",
  drop.refs = NULL,
  suffix = "_nrt"
)
}
\arguments{
\item{smru_ssm}{SSM-appended SMRU table file - output of \code{append_ssm}}

\item{fit}{final \code{foieGras} fit object}

\item{meta}{metadata}

\item{path}{path to write .csv files}

\item{drop.refs}{individual ids to be dropped}

\item{suffix}{suffix to add to .csv files (_nrt or _dm)}
}
\description{
reconfigure annotated tables - subsample predicted locations to 6-h interval, write to .csv and zip by campaign id
}
\examples{

}
