% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_meta.R
\name{clean_meta}
\alias{clean_meta}
\title{clean metadata}
\usage{
clean_meta(cids, smru, drop.refs = NULL, file = NULL)
}
\arguments{
\item{cids}{SMRU campaign ids}

\item{smru}{SMRU table}

\item{drop.refs}{SMRU refs to be dropped}

\item{file}{path to metadata .csv file}
}
\description{
restructures metadata, formats dates,
standardizes variable names, subsets to specified campaign ids, appends
dive start and end datetimes (for QC), & fills in missing
release_longitude/latitude's with the first best quality (lc 3,2, or 1)
Argos location after release_datetime
}
\examples{

}
