% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{run QC diagnostics}
\usage{
diagnostics(fit, fit1, diag, smru_ssm, meta, mpath = NULL, tpath = NULL)
}
\arguments{
\item{fit}{the final foieGras fit object from QC process}

\item{fit1}{the initial foieGras fit object from QC process}

\item{diag}{the standardized SMRU diag file (prior to truncation by metadata CTD start and end dates)}

\item{smru_ssm}{the ssm-annotated SMRU tables}

\item{meta}{metadata}

\item{mpath}{path to write map file}

\item{tpath}{path to write diagnostic table files}
}
\description{
produces a map of all QC''d tracks and generates diagnostic tables to assess QC run
}
\examples{

}
