% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{run QC diagnostics}
\usage{
diagnostics(
  fit,
  fit1,
  what = "p",
  cut,
  diag,
  smru_ssm,
  meta,
  mpath = NULL,
  tpath = NULL,
  QCmode = "nrt",
  ...
)
}
\arguments{
\item{fit}{the final aniMotum fit object from QC process}

\item{fit1}{the initial aniMotum fit object from QC process}

\item{what}{the SSM-estimated or rerouted locations to be used}

\item{cut}{logical; should predicted locations be dropped if keep = FALSE - ie. in a large data gap}

\item{diag}{the standardized SMRU diag file (prior to truncation by metadata CTD start and end dates)}

\item{smru_ssm}{the ssm-annotated SMRU tables}

\item{meta}{metadata}

\item{mpath}{path to write map file}

\item{tpath}{path to write diagnostic table files}

\item{QCmode}{specify whether QC is near real-time (nrt) or delayed-mode (dm),
in latter case start end end of dive data are displayed rather than ctd data}

\item{...}{extra arguments for aniMotum::map - used to generate maps}
}
\description{
produces a map of all QC''d tracks and generates diagnostic tables to assess QC run
}
