% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download SMRU or Wildlife Computers tag data from arbitrary sources}
\usage{
download_data(
  path = NULL,
  dest = NULL,
  source = "smru",
  unzip = FALSE,
  cids = NULL,
  user = NULL,
  pwd = NULL,
  wc.akey = "VVeerW+G6YUe7olzlrOr6q5o2Nkjx5PTEwuwElcPyKc=",
  wc.skey = "7k9MupziDacYNur/3IPMDjn7wum6oQk5eV2LRj86iDw=",
  ...
)
}
\arguments{
\item{path}{a url or local filepath pointing to a zipfile of tag datafiles}

\item{dest}{destination path to save download}

\item{source}{source type of data to be downloaded. Can be one of:
\itemize{
\item \code{smru} - SMRU Data Server;
\item \code{wc} - Wildlife Computers Data Portal API;
\item \code{googledrive} - a url link to a zipfile shared on a GoogleDrive;
\item \code{dropbox} - a url link to a zipfile shared on Dropbox;
\item \code{local} - a local filepath to a zipfile containing SMRU or WC tag datafiles.
}}

\item{unzip}{(logical) should the downloaded zipfile be unzipped.}

\item{cids}{SMRU tag deployment campaign id(s) to download, eg. "ct180"}

\item{user}{SMRU data server username as a quoted string}

\item{pwd}{SMRU data server password as a quoted string}

\item{wc.akey}{an Access Key issued by Wildlife Computers for their API}

\item{wc.skey}{a Secret Key issued by Wildlife Computers for their API}

\item{...}{additional arguments passed to \code{get_smru_mdb} or \code{get_wc_files()}}
}
\description{
Satellite tracking data, in a zipfile, are downloaded from a
user-specified URL (for accessing zipfiles on GoogleDrive or Dropbox),
accessed from the SMRU data server, or accessed from the Wildlife Computers
Portal API via the \code{source} argument.
}
