% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Generic wrapper to download data from arbitrary sources}
\usage{
get_data(
  url = NULL,
  dest = NULL,
  source = "smru",
  WC_AccessKey = "VVeerW+G6YUe7olzlrOr6q5o2Nkjx5PTEwuwElcPyKc=",
  WC_SecretKey = "7k9MupziDacYNur/3IPMDjn7wum6oQk5eV2LRj86iDw=",
  ...
)
}
\arguments{
\item{url}{url for data}

\item{dest}{destination path to save download}

\item{source}{source type of data to be downloaded. Can be one of:
`smru` - SMRU Data Server; `wc` - Wildlife Computers Data Portal API;
`googledrive` - a url link to a zipfile shared on a GoogleDrive;
`dropbox` - a url link to a zipfile shared on Dropbox;
`url` - a url link to a zipfile on a generic server.}

\item{WC_AccessKey}{an Access Key issued by Wildlife Computers for their API}

\item{WC_SecretKey}{a Secret Key issued by Wildlife Computers for their API}

\item{...}{additional arguments passed to `get_smru_mdb`}
}
\description{
SMRU SRDL-CTD data are downloaded from a user-specified URL
(which can include files on GoogleDrive or Dropbox), accessed from the SMRU
data server by calling `get_smru_mdb()`, or accessed from the Wildlife
Computers Portal API.
}
