% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{download deployment metadata}
\usage{
get_metadata(
  source = "smru",
  tag_mfr = "smru",
  tag_data = NULL,
  cid = NULL,
  user = NULL,
  pwd = NULL,
  dropIDs = NULL,
  file = NULL,
  meta.args,
  subset.ids = NULL,
  wc.meta = NULL
)
}
\arguments{
\item{source}{the source of the deployment metadata, current options are
\code{imos}, \code{smru}, or \code{atn}. If \code{source = 'imos'} or \code{source = 'atn'} then
metadata are obtained from user-provided .CSV file via the config.json file.
If \code{sourcce = 'smru'} then metadata are built from a combination of SMRU
server details & deployment details in the config.json file.}

\item{tag_mfr}{the tag manufacturer, current options are \code{smru} or \code{wc}
cid SMRU campaign ids}

\item{tag_data}{a list of either \code{smru} data tables or \code{wc} files as output by
\code{pull_data}.}

\item{cid}{SMRU campaign id must be provided when the tag_mfr is \code{smru}}

\item{user}{SMRU data server username as a quoted string - to be used only if
metadata are to be built from SMRU server details (\code{source = 'smru'}).}

\item{pwd}{SMRU data server password as a quoted string - to be used only if
metadata are to be built from SMRU server details (\code{source = 'smru'}).}

\item{dropIDs}{SMRU refs or WC ids to be dropped from QC}

\item{file}{path to metadata .csv file, if provided then metadata will be
read from the provided \code{source}}

\item{meta.args}{optional metadata fields to be passed from config file when
downloading tag metadata from SMRU server. Typically used only when no
metadata filepath is provided in the config file.}

\item{subset.ids}{a character vector of comma-separated (no spaces) WC UUID's
to be included in the QC. Ignored if \code{tag_mfr != "wc"}.}

\item{wc.meta}{an R data.frame of Wildlife Computers tag deployment metadata
obtained via \code{download_data()}. Typically used only when no metadata filepath
is provided in the config file.}
}
\description{
downloads, restructures & formats metadata, appends
dive/CTD start and end datetimes (for QC), & fills in missing required
metadata - eg. release_datetime, release_longitude/latitude's with data from
the GPS (if present) or Argos location file.
}
