% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{download deployment metadata}
\usage{
get_metadata(
  source = "smru",
  tag_mfr = "smru",
  tag_data = NULL,
  cids = NULL,
  dropIDs = NULL,
  file = NULL,
  enc = "UTF-8"
)
}
\arguments{
\item{source}{the source of the metadata, current options are \code{imos}, \code{smru},
\code{atn}.}

\item{tag_mfr}{the tag manufacturer, current options are \code{smru} or \code{wc}
cids SMRU campaign ids}

\item{tag_data}{a list of either \code{smru} data tables or \code{wc} files as output by
\code{pull_data}.}

\item{cids}{SMRU campaign id(s) must be provided when the tag_mfr is \code{smru}}

\item{dropIDs}{SMRU refs or WC ids to be dropped}

\item{file}{path to metadata .csv file, if provided then metadata will be
read from the provided \code{source}}

\item{enc}{set locale encoding to handle special characters; default is "UTF-8"}
}
\description{
downloads, restructures & formats metadata, appends
dive/CTD start and end datetimes (for QC), & fills in missing required
metadata - eg. release_datetime, release_longitude/latitude's with data from
the GPS (if present) or Argos location file.
}
