% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_filter.R
\name{multi_filter}
\alias{multi_filter}
\title{multi-filter}
\usage{
multi_filter(diag_sf, meta)
}
\arguments{
\item{diag_sf}{\code{sf}-projected diag to be used}

\item{meta}{metadata to split diag by species}
}
\description{
apply SSM filter to diag data across multiple processors
}
\examples{

}
