% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_filter.R
\name{multi_filter}
\alias{multi_filter}
\title{multi-filter}
\usage{
multi_filter(x, vmax = 4, model = "rw", ts = 6, verbose = FALSE)
}
\arguments{
\item{x}{\code{sf}-projected diag file of locations to be filtered}

\item{vmax}{for prefilter}

\item{model}{\code{aniMotum} model ("rw" orr "crw)}

\item{ts}{\code{foieGrsa} time.step}

\item{verbose}{turn on/off furrr::future_map progress indicator}
}
\description{
apply SSM filter to diag data across multiple processors
}
