% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull SMRU or WC data from downloaded data files}
\usage{
pull_data(path2data, source = "smru", cids = NULL, tag_mfr = "smru", ...)
}
\arguments{
\item{path2data}{path to data file(s)}

\item{source}{one of "smru", "wc", or "local"}

\item{cids}{SMRU campaign ids. If not specified then the cids are built from
the directory or filenames present in the \code{path2data} directory.}

\item{tag_mfr}{either "smru or "wc", ignored if \code{source} != "local"}

\item{...}{additional arguments passed to \code{pull_smru_tables}}
}
\description{
extracts specified data files from SMRU .mdb files or WC zipfiles.
For WC data files, extracts data from \code{-Locations.csv}, \code{-FastGPS.csv},
\code{Histos.csv}, \code{MinMaxDepth.csv}, \code{HaulOut.csv}, and \code{SST.csv} files.
Extracted data are aggregated across individual tags and returned in a
single named list with the following data.frames:
\itemize{
\item argos
\item fastgps
\item histos
\item depth
\item haulout
\item sst
WC tag data files downloaded via \code{download_data} will be stored in separate,
tag-specific subdirectories. \code{path2data} should point to the outer directory.
The function always chooses the latest WC-processed \code{-Locations.csv} &
\code{FastGPS.csv} files.
}
}
