% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_wc_data.R
\name{pull_wc_data}
\alias{pull_wc_data}
\title{Pull data from Wildlife Computers tag data files}
\usage{
pull_wc_data(path2data)
}
\arguments{
\item{path2data}{path to all WC tag data files.}
}
\description{
extracts data from \code{X-Locations.csv} (Argos), \code{X-FastGPS.csv},
\code{ECDHistos.csv}, \code{Histos.csv}, \code{MixLayer.csv}, \code{PDTs.csv}, \code{DSA.csv},
\code{MinMaxDepth.csv}, \code{HaulOut.csv}, and \code{SST.csv} files.
Extracted data are aggregated across individual tags and returned in a
single named list with the following data.frames:
\itemize{
\item Argos
\item FastGPS
\item ECDHistos_SCOUT_TEMP_361A
\item ECDHistos_SCOUT_DSA
\item Histos
\item Mixlayer
\item PDTs
\item DSA
\item MinMaxDepth
\item Haulout
\item SST
WC tag data files downloaded via \code{download_data} will be stored in separate,
tag-specific subdirectories. \code{path2data} should point to the outer directory.
}
}
