% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_multi_filter.R
\name{redo_multi_filter}
\alias{redo_multi_filter}
\title{redo failed multi-filter cases}
\usage{
redo_multi_filter(fit, diag_sf, model = "crw", ts = 3, min.dt = 180,
  map = NULL)
}
\arguments{
\item{fit}{foieGras fit object from first round of filtering}

\item{diag_sf}{\code{sf}-projected diag to be used}

\item{model}{model argument ("rw" or "crw) for \code{foieGras::fit_ssm}}

\item{ts}{time.step argument for \code{foieGras::fit_ssm}}

\item{min.dt}{min.dt argument for \code{foieGras::fit_ssm}}

\item{map}{params to fix}
}
\description{
re-apply SSM filter to diag data for id's that failed to converge. parallelized
}
\examples{

}
