% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_multi_filter.R
\name{redo_multi_filter}
\alias{redo_multi_filter}
\title{redo failed multi-filter cases}
\usage{
redo_multi_filter(
  fit,
  diag_sf,
  model = "crw",
  ts = 3,
  vmax = 2,
  ang = c(15, 25),
  distlim = c(1500, 5000),
  min.dt = 180,
  map = NULL,
  reroute = TRUE,
  dist = 1500,
  barrier = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fit}{aniMotum fit object from first round of filtering}

\item{diag_sf}{\code{sf}-projected diag to be used}

\item{model}{model argument ("rw" or "crw) for \code{aniMotum::fit_ssm}}

\item{ts}{time.step argument for \code{aniMotum::fit_ssm}}

\item{vmax}{threshold travel speed (m/s) to apply during track pre-filtering}

\item{ang}{sdafilter argument}

\item{distlim}{sdafilter argument}

\item{min.dt}{min.dt argument for \code{aniMotum::fit_ssm}}

\item{map}{params to fix}

\item{reroute}{(logical) should SSM-predicted locations be re-routed off of
land (default is TRUE)}

\item{dist}{the distance (in km) to buffer around predicted locations. This
buffer allows a larger portion of coastline to be selected for rerouting any
locations that are on land. More coastline polygon data can help rerouting, but
too much will make computation very slow.}

\item{barrier}{add a custom POLYGON/MULTIPOLYGON shapefile to use as a land
barrier. Default (NULL) reverts to the \code{aniMotum} default land barrier.}

\item{verbose}{turn on/off furrr::future_map progress indicator}

\item{...}{additional arguments to \code{aniMotum::route_path}}
}
\description{
re-apply SSM filter to diag data for id's that failed to converge. parallelized
}
