% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_write_csv.R
\name{smru_write_csv}
\alias{smru_write_csv}
\title{reconfigure annotated SMRU tables, test for expected data schema for AODN}
\usage{
smru_write_csv(
  smru_ssm,
  fit,
  what,
  meta,
  program = "imos",
  proj = NULL,
  test = TRUE,
  path = NULL,
  dropIDs = NULL,
  suffix = "_nrt"
)
}
\arguments{
\item{smru_ssm}{SSM-appended SMRU table file - output of \code{append_ssm}}

\item{fit}{final \code{aniMotum} fit object}

\item{what}{specify whether predicted or rerouted locations are to be used}

\item{meta}{metadata}

\item{program}{Determines structure of output metadata. The \code{imos}, \code{otn}, & \code{atn} programs
have their own defined metadata structures, any other program name is treated as "Generic" but may not work properly.}

\item{proj}{the proj4string specified in the .JSON config file & used to project
the location data prior to SSM fitting. It is passed in here to be added to the
output metadata .CSV file}

\item{test}{should variables be tested for standards compliance, default is TRUE.
Standards compliance is specific to the program. Currently, only program = \code{imos}
has defined variable standard against which output compliance is tested.}

\item{path}{path to write .csv files}

\item{dropIDs}{individual SMRU ids to be dropped}

\item{suffix}{suffix to add to .csv files (_nrt, _dm, or _hist)}
}
\description{
reconfigure annotated tables - subsample predicted locations to 6-h interval, write to .csv and zip by campaign id
}
