% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smru_write_csv.R
\name{smru_write_ssm}
\alias{smru_write_ssm}
\title{ssmoutputs table write to .csv}
\usage{
smru_write_ssm(
  locs_out,
  meta,
  program = "imos",
  path = NULL,
  dropIDs = NULL,
  suffix = "_nrt"
)
}
\arguments{
\item{locs_out}{SSM-predicted locations}

\item{meta}{metadata}

\item{program}{Determines structure of output metadata. Currently, either \code{imos} or \code{atn}.}

\item{path}{path to write .csv files}

\item{dropIDs}{individual ids to be dropped}

\item{suffix}{suffix to add to .csv files (_nrt, _dm, or _hist)}
}
\description{
write to .csv - format depends on program (IMOS, ATN)
}
\keyword{internal}
