% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_diag_sf.R
\name{truncate_diag_sf}
\alias{truncate_diag_sf}
\title{truncate diag data and convert to sf tibble}
\usage{
truncate_diag_sf(diag, meta, crs = "+init=epsg:3395 +units=km")
}
\arguments{
\item{diag}{diag to use}

\item{meta}{metadata used to truncate start of diag data for each individual}

\item{crs}{a proj4string or EPSG code to re-project diag locations from longlat}
}
\description{
truncate start of diag data for each individual, using deployment date-time from metadata
}
\examples{

}
