% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_get_files.R
\name{wc_get_files}
\alias{wc_get_files}
\title{WC API access to tag data files}
\usage{
wc_get_files(
  dest = NULL,
  a.key = NULL,
  s.key = NULL,
  owner.id = NULL,
  collaborator = TRUE,
  unzip = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{a.key}{an Access Key issued by Wildlife Computers for their API}

\item{s.key}{a Secret Key issued by Wildlife Computers for their API}

\item{owner.id}{a WC data owner ID}

\item{collaborator}{(logical) should data owned by collaborators be
downloaded. Ignored if \code{owner.id} provided.}

\item{unzip}{(logical) should deployment zipfile be unzipped into destination
directory.}
}
\description{
Download satellite tracking data and associated tag files via
the Wildlife Computers Portal API. User-supplied WC Access Key, WC Secret Key,

Satellite tracking data are downloaded from a user-specified URL
(which can include files on GoogleDrive or Dropbox), accessed from the SMRU
data server by calling \code{get_smru_mdb()}, or accessed from the Wildlife
Computers Portal API by calling \code{get_wc_files()}.
}
\keyword{internal}
